requirements.txt
anyconfig
colorama==0.3.7
docopt==0.6.2
docker-py
Jinja2>=2.8
libvirt-python
paramiko
pbr
pexpect==4.1.0
ptyprocess
pytest-xdist
python-vagrant==0.5.14
PyYAML==3.11
pytz
sh==1.11
tabulate==0.7.5
testinfra
shade
